requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.2
dj-database-url>=2.2
whitenoise>=6.7
gunicorn>=22.0
psycopg2-binary>=2.9
twilio>=9.0
